+++ CHANGES.rst
@@ -6,6 +6,8 @@ New Features
 
 - Datasets can now be renamed via API and UI through the viewer data-menu. [#3960, #3968]
 
+- Allow creating new viewers (in deconfigged) when adding results from plugins. [#3966]
+
 Cubeviz
 ^^^^^^^